Add notebooks with solutions
These are solutions to "Scientific Computing with Python" from freeCodeCamp.org, written in Jupyter notebook. Language is python. Environment file is included.
There are separate notebooks with descriptions taken from the project, and separate with solutions, this is because I find it easier to work when I have them on separate screens. <br><br>
Project list from certification:
* Arithmetic Formatter
* Time Calculator
* Budget App
* Polygon Area Calculator
* Probability Calculator